wasm: close base64 encoder
Close the base64 encoder before returning the buffer. This ensures that
all the bytes are written to the buffer. Without this, the last few
bytes may be missing, resulting in a corrupted image. Usually, this
would only corrupt the checksum, and the image would still be displayed
fine. However, some applications do not like this (e.g., slack).